Organize better all the modules
