+++ rvc/configs/v1/40000.json
@@ -37,7 +37,7 @@
     "n_layers": 6,
     "kernel_size": 3,
     "p_dropout": 0,
-    "resblock": "1",
+    "resblock": 1,
     "resblock_kernel_sizes": [3, 7, 11],
     "resblock_dilation_sizes": [
       [1, 3, 5],
