+++ rvc/lib/algorithm/discriminators/sub/scale.py
@@ -2,7 +2,7 @@ import torch
 import torch.nn as nn
 import torch.nn.functional as F
 from torch.nn.utils.parametrizations import spectral_norm, weight_norm
-from infer.lib.infer_pack.san_modules import SANConv1d
+from rvc.lib.algorithm.layers.san import SANConv1d
 
 from rvc.lib.algorithm.residuals import LRELU_SLOPE
