+++ rvc/lib/algorithm/generators/bigvgan/bigvgan.py
@@ -1,0 +1,216 @@
+import torch
+import torch.nn as nn
+import torch.nn.functional as F
+import numpy as np
+
+from torch.nn import Conv1d
+from torch.nn import ConvTranspose1d
+from torch.nn.utils.parametrizations import weight_norm
+from torch.nn.utils import remove_weight_norm
+
+import sys
+import os
+
+now_dir = os.getcwd()
+sys.path.append(now_dir)
+
+from rvc.lib.algorithm.alias.act import SnakeAlias
+from rvc.lib.algorithm.commons import init_weights
+from rvc.lib.algorithm.layers.amp import AMPBlock
+from rvc.lib.algorithm.generators.bigvgan.nsf import SourceModuleHnNSF
+
+
+class SpeakerAdapter(nn.Module):
+    def __init__(self, speaker_dim, adapter_dim, epsilon=1e-5):
+        super(SpeakerAdapter, self).__init__()
+        self.speaker_dim = speaker_dim
+        self.adapter_dim = adapter_dim
+        self.epsilon = epsilon
+        self.W_scale = nn.Linear(self.speaker_dim, self.adapter_dim)
+        self.W_bias = nn.Linear(self.speaker_dim, self.adapter_dim)
+        self.reset_parameters()
+
+    def reset_parameters(self):
+        torch.nn.init.constant_(self.W_scale.weight, 0.0)
+        torch.nn.init.constant_(self.W_scale.bias, 1.0)
+        torch.nn.init.constant_(self.W_bias.weight, 0.0)
+        torch.nn.init.constant_(self.W_bias.bias, 0.0)
+
+    def forward(self, x, speaker_embedding):
+        x = x.transpose(1, -1)
+        mean = x.mean(dim=-1, keepdim=True)
+        var = ((x - mean) ** 2).mean(dim=-1, keepdim=True)
+        std = (var + self.epsilon).sqrt()
+        y = (x - mean) / std
+        scale = self.W_scale(speaker_embedding)
+        bias = self.W_bias(speaker_embedding)
+        y *= scale.unsqueeze(1)
+        y += bias.unsqueeze(1)
+        y = y.transpose(1, -1)
+        return y
+
+
+class BigVGAN(torch.nn.Module):
+    # this is our main BigVGAN model. Applies anti-aliased periodic activation for resblocks.
+    def __init__(
+        self,
+        resblock_kernel_sizes,
+        resblock_dilation_sizes,
+        upsample_rates,
+        upsample_kernel_sizes,
+        upsample_input,
+        upsample_initial_channel,
+        sample_rate,
+        spk_dim,
+    ):
+        super(BigVGAN, self).__init__()
+        self.num_kernels = len(resblock_kernel_sizes)
+        self.num_upsamples = len(upsample_rates)
+        self.adapter = SpeakerAdapter(spk_dim, upsample_input)
+        self.conv_pre = Conv1d(
+            upsample_input, upsample_initial_channel, 7, 1, padding=3
+        )
+
+        self.f0_upsamp = torch.nn.Upsample(scale_factor=np.prod(upsample_rates))
+        self.m_source = SourceModuleHnNSF(sample_rate=sample_rate, harmonic_num=10)
+        self.noise_convs = nn.ModuleList()
+        # transposed conv-based upsamplers. does not apply anti-aliasing
+        self.ups = nn.ModuleList()
+        for i, (u, k) in enumerate(zip(upsample_rates, upsample_kernel_sizes)):
+            # print(f'ups: {i} {k}, {u}, {(k - u) // 2}')
+            # base
+            self.ups.append(
+                weight_norm(
+                    ConvTranspose1d(
+                        upsample_initial_channel // (2**i),
+                        upsample_initial_channel // (2 ** (i + 1)),
+                        k,
+                        u,
+                        padding=(k - u) // 2,
+                    )
+                )
+            )
+            # nsf
+            if i + 1 < len(upsample_rates):
+                stride_f0 = np.prod(upsample_rates[i + 1 :])
+                stride_f0 = int(stride_f0)
+                self.noise_convs.append(
+                    Conv1d(
+                        1,
+                        upsample_initial_channel // (2 ** (i + 1)),
+                        kernel_size=stride_f0 * 2,
+                        stride=stride_f0,
+                        padding=stride_f0 // 2,
+                    ).half()
+                )
+            else:
+                self.noise_convs.append(
+                    Conv1d(
+                        1, upsample_initial_channel // (2 ** (i + 1)), kernel_size=1
+                    ).half()
+                )
+
+        # residual blocks using anti-aliased multi-periodicity composition modules (AMP)
+        self.resblocks = nn.ModuleList()
+        for i in range(len(self.ups)):
+            ch = upsample_initial_channel // (2 ** (i + 1))
+            for k, d in zip(resblock_kernel_sizes, resblock_dilation_sizes):
+                self.resblocks.append(AMPBlock(ch, k, d))
+
+        # post conv
+        self.activation_post = SnakeAlias(ch)
+        self.conv_post = Conv1d(ch, 1, 7, 1, padding=3, bias=False)
+        # weight initialization
+        self.ups.apply(init_weights)
+
+    def forward(self, x, f0, g):
+        if g.size(-1) == 1:
+            speaker_embedding = g.squeeze(-1)
+        else:
+            speaker_embedding = g
+        # Perturbation
+        x = x + torch.randn_like(x)
+        # adapter
+        x = self.adapter(x, speaker_embedding=speaker_embedding)
+        x = self.conv_pre(x)
+        x = x * torch.tanh(F.softplus(x))
+        # nsf
+        f0 = f0[:, None]
+        f0 = self.f0_upsamp(f0).transpose(1, 2)
+        har_source = self.m_source(f0)
+        har_source = har_source.transpose(1, 2)
+
+        for i in range(self.num_upsamples):
+            # upsampling
+            x = self.ups[i](x)
+            # nsf
+            x_source = self.noise_convs[i](har_source.half())
+            x = x + x_source
+            # AMP blocks
+            xs = None
+            for j in range(self.num_kernels):
+                if xs is None:
+                    xs = self.resblocks[i * self.num_kernels + j](x)
+                else:
+                    xs += self.resblocks[i * self.num_kernels + j](x)
+            x = xs / self.num_kernels
+
+        # post conv
+        x = self.activation_post(x)
+        x = self.conv_post(x)
+        x = torch.tanh(x)
+        return x
+
+    def remove_weight_norm(self):
+        for l in self.ups:
+            remove_weight_norm(l)
+        for l in self.resblocks:
+            l.remove_weight_norm()
+
+    def eval(self, inference=False):
+        super(BigVGAN, self).eval()
+        # don't remove weight norm while validation in training loop
+        if inference:
+            self.remove_weight_norm()
+
+    def pitch2source(self, f0):
+        f0 = f0[:, None]
+        f0 = self.f0_upsamp(f0).transpose(1, 2)  # [1,len,1]
+        har_source = self.m_source(f0)
+        har_source = har_source.transpose(1, 2)  # [1,1,len]
+        return har_source
+
+    def source2wav(self, audio):
+        MAX_WAV_VALUE = 32768.0
+        audio = audio.squeeze()
+        audio = MAX_WAV_VALUE * audio
+        audio = audio.clamp(min=-MAX_WAV_VALUE, max=MAX_WAV_VALUE - 1)
+        audio = audio.short()
+        return audio.cpu().detach().numpy()
+
+    def inference(self, x, har_source, g):
+        # adapter
+        x = self.adapter(x, g)
+        x = self.conv_pre(x)
+        x = x * torch.tanh(F.softplus(x))
+
+        for i in range(self.num_upsamples):
+            # upsampling
+            x = self.ups[i](x)
+            # nsf
+            x_source = self.noise_convs[i](har_source)
+            x = x + x_source
+            # AMP blocks
+            xs = None
+            for j in range(self.num_kernels):
+                if xs is None:
+                    xs = self.resblocks[i * self.num_kernels + j](x)
+                else:
+                    xs += self.resblocks[i * self.num_kernels + j](x)
+            x = xs / self.num_kernels
+
+        # post conv
+        x = self.activation_post(x)
+        x = self.conv_post(x)
+        x = torch.tanh(x)
+        return x
