+++ rvc/lib/algorithm/generators/hifigan/hifigan.py
@@ -4,151 +4,18 @@ from torch.nn.utils import remove_weight_norm
 from torch.nn.utils.parametrizations import weight_norm
 from typing import Optional
 
+import sys
+import os
+
+now_dir = os.getcwd()
+sys.path.append(now_dir)
+
 from rvc.lib.algorithm.residuals import LRELU_SLOPE, ResBlock1, ResBlock2
 from rvc.lib.algorithm.commons import init_weights
+from rvc.lib.algorithm.generators.hifigan.nsf import SourceModuleHnNSF
 
 
-class SineGen(torch.nn.Module):
-    """Sine wave generator.
-
-    Args:
-        samp_rate (int): Sampling rate in Hz.
-        harmonic_num (int, optional): Number of harmonic overtones. Defaults to 0.
-        sine_amp (float, optional): Amplitude of sine waveform. Defaults to 0.1.
-        noise_std (float, optional): Standard deviation of Gaussian noise. Defaults to 0.003.
-        voiced_threshold (float, optional): F0 threshold for voiced/unvoiced classification. Defaults to 0.
-        flag_for_pulse (bool, optional): Whether this SineGen is used inside PulseGen. Defaults to False.
-    """
-
-    def __init__(
-        self,
-        samp_rate,
-        harmonic_num=0,
-        sine_amp=0.1,
-        noise_std=0.003,
-        voiced_threshold=0,
-        flag_for_pulse=False,
-    ):
-        super(SineGen, self).__init__()
-        self.sine_amp = sine_amp
-        self.noise_std = noise_std
-        self.harmonic_num = harmonic_num
-        self.dim = self.harmonic_num + 1
-        self.sample_rate = samp_rate
-        self.voiced_threshold = voiced_threshold
-
-    def _f02uv(self, f0):
-        """Converts F0 to voiced/unvoiced signal.
-
-        Args:
-            f0 (torch.Tensor): F0 tensor with shape (batch_size, length, 1).
-        """
-        # generate uv signal
-        uv = torch.ones_like(f0)
-        uv = uv * (f0 > self.voiced_threshold)
-        return uv
-
-    def forward(self, f0: torch.Tensor, upp: int):
-        """Generates sine waves.
-
-        Args:
-            f0 (torch.Tensor): F0 tensor with shape (batch_size, length, 1).
-            upp (int): Upsampling factor.
-        """
-        with torch.no_grad():
-            f0 = f0[:, None].transpose(1, 2)
-            f0_buf = torch.zeros(f0.shape[0], f0.shape[1], self.dim, device=f0.device)
-            # fundamental component
-            f0_buf[:, :, 0] = f0[:, :, 0]
-            for idx in range(self.harmonic_num):
-                f0_buf[:, :, idx + 1] = f0_buf[:, :, 0] * (
-                    idx + 2
-                )  # idx + 2: the (idx+1)-th overtone, (idx+2)-th harmonic
-            rad_values = (f0_buf / float(self.sample_rate)) % 1
-            rand_ini = torch.rand(
-                f0_buf.shape[0], f0_buf.shape[2], device=f0_buf.device
-            )
-            rand_ini[:, 0] = 0
-            rad_values[:, 0, :] = rad_values[:, 0, :] + rand_ini
-            tmp_over_one = torch.cumsum(rad_values, 1)
-            tmp_over_one *= upp
-            tmp_over_one = torch.nn.functional.interpolate(
-                tmp_over_one.transpose(2, 1),
-                scale_factor=float(upp),
-                mode="linear",
-                align_corners=True,
-            ).transpose(2, 1)
-            rad_values = torch.nn.functional.interpolate(
-                rad_values.transpose(2, 1), scale_factor=float(upp), mode="nearest"
-            ).transpose(2, 1)
-            tmp_over_one %= 1
-            tmp_over_one_idx = (tmp_over_one[:, 1:, :] - tmp_over_one[:, :-1, :]) < 0
-            cumsum_shift = torch.zeros_like(rad_values)
-            cumsum_shift[:, 1:, :] = tmp_over_one_idx * -1.0
-            sine_waves = torch.sin(
-                torch.cumsum(rad_values + cumsum_shift, dim=1) * 2 * torch.pi
-            )
-            sine_waves = sine_waves * self.sine_amp
-            uv = self._f02uv(f0)
-            uv = torch.nn.functional.interpolate(
-                uv.transpose(2, 1), scale_factor=float(upp), mode="nearest"
-            ).transpose(2, 1)
-            noise_amp = uv * self.noise_std + (1 - uv) * self.sine_amp / 3
-            noise = noise_amp * torch.randn_like(sine_waves)
-            sine_waves = sine_waves * uv + noise
-        return sine_waves, uv, noise
-
-
-class SourceModuleHnNSF(torch.nn.Module):
-    """Source Module for harmonic-plus-noise excitation.
-
-    Args:
-        sample_rate (int): Sampling rate in Hz.
-        harmonic_num (int, optional): Number of harmonics above F0. Defaults to 0.
-        sine_amp (float, optional): Amplitude of sine source signal. Defaults to 0.1.
-        add_noise_std (float, optional): Standard deviation of additive Gaussian noise. Defaults to 0.003.
-        voiced_threshod (float, optional): Threshold to set voiced/unvoiced given F0. Defaults to 0.
-        is_half (bool, optional): Whether to use half precision. Defaults to True.
-    """
-
-    def __init__(
-        self,
-        sample_rate,
-        harmonic_num=0,
-        sine_amp=0.1,
-        add_noise_std=0.003,
-        voiced_threshod=0,
-        is_half=True,
-    ):
-        super(SourceModuleHnNSF, self).__init__()
-
-        self.sine_amp = sine_amp
-        self.noise_std = add_noise_std
-        self.is_half = is_half
-        # to produce sine waveforms
-        self.l_sin_gen = SineGen(
-            sample_rate, harmonic_num, sine_amp, add_noise_std, voiced_threshod
-        )
-
-        # to merge source harmonics into a single excitation
-        self.l_linear = torch.nn.Linear(harmonic_num + 1, 1)
-        self.l_tanh = torch.nn.Tanh()
-        # self.ddtype:int = -1
-
-    def forward(self, x: torch.Tensor, upp: int = 1):
-        """Generates the source excitation.
-
-        Args:
-            x (torch.Tensor): F0 tensor with shape (batch_size, length, 1).
-            upp (int): Upsampling factor.
-        """
-        sine_wavs, uv, _ = self.l_sin_gen(x, upp)
-        sine_wavs = sine_wavs.to(dtype=self.l_linear.weight.dtype)
-        sine_merge = self.l_tanh(self.l_linear(sine_wavs))
-        return sine_merge, None, None  # noise, uv
-
-
-class GeneratorNSF_HIFIGAN(torch.nn.Module):
+class HiFiGAN(torch.nn.Module):
     """Generator for synthesizing audio using the NSF (Neural Source Filter) approach.
 
     Args:
@@ -177,7 +44,7 @@ class GeneratorNSF_HIFIGAN(torch.nn.Module):
         sr,
         is_half=False,
     ):
-        super(GeneratorNSF_HIFIGAN, self).__init__()
+        super(HiFiGAN, self).__init__()
         self.num_kernels = len(resblock_kernel_sizes)
         self.num_upsamples = len(upsample_rates)
 
@@ -189,7 +56,7 @@ class GeneratorNSF_HIFIGAN(torch.nn.Module):
         self.conv_pre = torch.nn.Conv1d(
             initial_channel, upsample_initial_channel, 7, 1, padding=3
         )
-        resblock = ResBlock1 if resblock == "1" else ResBlock2
+        resblock = ResBlock1 if resblock == 1 else ResBlock2
 
         self.ups = torch.nn.ModuleList()
         for i, (u, k) in enumerate(zip(upsample_rates, upsample_kernel_sizes)):
@@ -239,7 +106,7 @@ class GeneratorNSF_HIFIGAN(torch.nn.Module):
 
     def forward(self, x, f0, g: Optional[torch.Tensor] = None):
         """Generates audio using the NSF approach.
-        
+
         Args:
             x (torch.Tensor): Input tensor with shape (batch_size, initial_channel, length).
             f0 (torch.Tensor): F0 tensor with shape (batch_size, length, 1).
