+++ rvc/lib/algorithm/modules.py
@@ -1,6 +1,13 @@
+import os
+import sys
 import torch
-from rvc.lib.algorithm.commons import fused_add_tanh_sigmoid_multiply
 from typing import Optional
+
+now_dir = os.getcwd()
+sys.path.append(now_dir)
+
+from rvc.lib.algorithm.commons import fused_add_tanh_sigmoid_multiply
+
 
 class WaveNet(torch.nn.Module):
     """WaveNet residual blocks as used in WaveGlow
@@ -41,10 +48,12 @@ class WaveNet(torch.nn.Module):
             cond_layer = torch.nn.Conv1d(
                 gin_channels, 2 * hidden_channels * n_layers, 1
             )
-            self.cond_layer = torch.nn.utils.parametrizations.weight_norm(cond_layer, name="weight")
+            self.cond_layer = torch.nn.utils.parametrizations.weight_norm(
+                cond_layer, name="weight"
+            )
 
         for i in range(n_layers):
-            dilation = dilation_rate ** i
+            dilation = dilation_rate**i
             padding = int((kernel_size * dilation - dilation) / 2)
             in_layer = torch.nn.Conv1d(
                 hidden_channels,
@@ -53,7 +62,9 @@ class WaveNet(torch.nn.Module):
                 dilation=dilation,
                 padding=padding,
             )
-            in_layer = torch.nn.utils.parametrizations.weight_norm(in_layer, name="weight")
+            in_layer = torch.nn.utils.parametrizations.weight_norm(
+                in_layer, name="weight"
+            )
             self.in_layers.append(in_layer)
 
             # last one is not necessary
@@ -63,11 +74,13 @@ class WaveNet(torch.nn.Module):
                 res_skip_channels = hidden_channels
 
             res_skip_layer = torch.nn.Conv1d(hidden_channels, res_skip_channels, 1)
-            res_skip_layer = torch.nn.utils.parametrizations.weight_norm(res_skip_layer, name="weight")
+            res_skip_layer = torch.nn.utils.parametrizations.weight_norm(
+                res_skip_layer, name="weight"
+            )
             self.res_skip_layers.append(res_skip_layer)
 
     def forward(
-            self, x: torch.Tensor, x_mask: torch.Tensor, g: Optional[torch.Tensor] = None
+        self, x: torch.Tensor, x_mask: torch.Tensor, g: Optional[torch.Tensor] = None
     ):
         """
         Perform inference of a model.
@@ -86,12 +99,12 @@ class WaveNet(torch.nn.Module):
             g = self.cond_layer(g)
 
         for i, (in_layer, res_skip_layer) in enumerate(
-                zip(self.in_layers, self.res_skip_layers)
+            zip(self.in_layers, self.res_skip_layers)
         ):
             x_in = in_layer(x)
             if g is not None:
                 cond_offset = i * 2 * self.hidden_channels
-                g_l = g[:, cond_offset: cond_offset + 2 * self.hidden_channels, :]
+                g_l = g[:, cond_offset : cond_offset + 2 * self.hidden_channels, :]
             else:
                 g_l = torch.zeros_like(x_in)
 
@@ -102,7 +115,7 @@ class WaveNet(torch.nn.Module):
             if i < self.n_layers - 1:
                 res_acts = res_skip_acts[:, : self.hidden_channels, :]
                 x = (x + res_acts) * x_mask
-                output = output + res_skip_acts[:, self.hidden_channels:, :]
+                output = output + res_skip_acts[:, self.hidden_channels :, :]
             else:
                 output = output + res_skip_acts
         return output * x_mask
@@ -114,27 +127,27 @@ class WaveNet(torch.nn.Module):
             torch.nn.utils.remove_weight_norm(l)
         for l in self.res_skip_layers:
             torch.nn.utils.remove_weight_norm(l)
+
     def __prepare_scriptable__(self):
         if self.gin_channels != 0:
             for hook in self.cond_layer._forward_pre_hooks.values():
                 if (
-                        hook.__module__ == "torch.nn.utils.parametrizations.weight_norm"
-                        and hook.__class__.__name__ == "WeightNorm"
+                    hook.__module__ == "torch.nn.utils.parametrizations.weight_norm"
+                    and hook.__class__.__name__ == "WeightNorm"
                 ):
                     torch.nn.utils.remove_weight_norm(self.cond_layer)
         for l in self.in_layers:
             for hook in l._forward_pre_hooks.values():
                 if (
-                        hook.__module__ == "torch.nn.utils.parametrizations.weight_norm"
-                        and hook.__class__.__name__ == "WeightNorm"
+                    hook.__module__ == "torch.nn.utils.parametrizations.weight_norm"
+                    and hook.__class__.__name__ == "WeightNorm"
                 ):
                     torch.nn.utils.remove_weight_norm(l)
         for l in self.res_skip_layers:
             for hook in l._forward_pre_hooks.values():
                 if (
-                        hook.__module__ == "torch.nn.utils.parametrizations.weight_norm"
-                        and hook.__class__.__name__ == "WeightNorm"
+                    hook.__module__ == "torch.nn.utils.parametrizations.weight_norm"
+                    and hook.__class__.__name__ == "WeightNorm"
                 ):
                     torch.nn.utils.remove_weight_norm(l)
         return self
-
