+++ rvc/lib/algorithm/normalization.py
@@ -1,7 +1,5 @@
 import torch
 
-
-
 class LayerNorm(torch.nn.Module):
     """Layer normalization module.
 
@@ -25,5 +23,7 @@ class LayerNorm(torch.nn.Module):
             x (torch.Tensor): Input tensor of shape (batch_size, channels, time_steps).
         """
         x = x.transpose(1, -1)
-        x = torch.nn.functional.layer_norm(x, (self.channels,), self.gamma, self.beta, self.eps)
+        x = torch.nn.functional.layer_norm(
+            x, (self.channels,), self.gamma, self.beta, self.eps
+        )
         return x.transpose(1, -1)
